correction base_url de l'api de liste des libellés pour être sur le bon µS (routage de l'api gw)
