--- operations/src/main/java/io/github/vzwingma/finances/budget/serverless/services/operations/api/OperationsRessource.java
@@ -1,86 +1,0 @@
-package io.github.vzwingma.finances.budget.serverless.services.operations.api;
-
-import io.github.vzwingma.finances.budget.serverless.services.operations.api.enums.OperationsAPIEnum;
-import io.github.vzwingma.finances.budget.serverless.services.operations.business.model.budget.BudgetMensuel;
-import io.github.vzwingma.finances.budget.serverless.services.operations.business.ports.IBudgetAppProvider;
-import io.github.vzwingma.finances.budget.services.communs.api.AbstractAPIInterceptors;
-import io.github.vzwingma.finances.budget.services.communs.data.trace.BusinessTraceContext;
-import io.github.vzwingma.finances.budget.services.communs.data.trace.BusinessTraceContextKeyEnum;
-import io.github.vzwingma.finances.budget.services.communs.utils.exceptions.BadParametersException;
-import io.smallrye.mutiny.Multi;
-import jakarta.annotation.security.RolesAllowed;
-import jakarta.inject.Inject;
-import jakarta.ws.rs.Consumes;
-import jakarta.ws.rs.GET;
-import jakarta.ws.rs.Path;
-import jakarta.ws.rs.Produces;
-import jakarta.ws.rs.container.ContainerRequestContext;
-import jakarta.ws.rs.container.ContainerResponseContext;
-import jakarta.ws.rs.core.MediaType;
-import org.eclipse.microprofile.openapi.annotations.Operation;
-import org.eclipse.microprofile.openapi.annotations.media.Content;
-import org.eclipse.microprofile.openapi.annotations.media.Schema;
-import org.eclipse.microprofile.openapi.annotations.responses.APIResponse;
-import org.eclipse.microprofile.openapi.annotations.responses.APIResponses;
-import org.jboss.resteasy.reactive.RestPath;
-import org.jboss.resteasy.reactive.RestQuery;
-import org.jboss.resteasy.reactive.server.ServerRequestFilter;
-import org.jboss.resteasy.reactive.server.ServerResponseFilter;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-
-@Path(OperationsAPIEnum.COMPTE_BASE)
-public class OperationsRessource extends AbstractAPIInterceptors {
-
-
-    private static final Logger LOG = LoggerFactory.getLogger(OperationsRessource.class);
-
-
-    @Inject
-    IBudgetAppProvider budgetService;
-
-    /**
-     * Liste des libellés des opérations
-     * @param idCompte id du compte
-     * @return liste des libellés
-     */
-    @Operation(description="Liste des libellés des opérations")
-    @APIResponses(value = {
-            @APIResponse(responseCode = "200", description = "Liste des opérations",
-                    content = { @Content(mediaType = "application/json", schema = @Schema(implementation = BudgetMensuel.class))}),
-            @APIResponse(responseCode = "400", description = "Paramètres incorrects"),
-            @APIResponse(responseCode = "401", description = "Utilisateur non authentifié"),
-            @APIResponse(responseCode = "403", description = "Opération non autorisée"),
-            @APIResponse(responseCode = "404", description = "Données introuvables"),
-            @APIResponse(responseCode = "423", description = "Compte clos")
-    })
-    @GET
-    @Path(value= OperationsAPIEnum.OPERATIONS_LIBELLES)
-    @RolesAllowed({ OperationsAPIEnum.OPERATIONS_ROLE })
-    @Consumes(MediaType.APPLICATION_JSON)
-    @Produces(MediaType.APPLICATION_JSON)
-    public Multi<String> libellesOperationsCompte(@RestPath("idCompte") String idCompte) {
-
-        LOG.info("Libelles des opérations du Compte " + idCompte);
-
-        if(idCompte != null){
-            BusinessTraceContext.getclear().put(BusinessTraceContextKeyEnum.COMPTE, idCompte).put(BusinessTraceContextKeyEnum.USER, super.getAuthenticatedUser());
-            return budgetService.getLibellesOperations(idCompte, super.getAuthenticatedUser());
-        }
-        else {
-            return Multi.createFrom().failure(new BadParametersException("Le paramètre idCompte est obligatoire"));
-        }
-    }
-
-
-
-    @Override
-    @ServerRequestFilter(preMatching = true)
-    public void preMatchingFilter(ContainerRequestContext requestContext) {
-        super.preMatchingFilter(requestContext);
-    }
-    @Override
-    @ServerResponseFilter
-    public void postMatchingFilter(ContainerResponseContext responseContext) { super.postMatchingFilter(responseContext); }
-}
